Add files via upload
The sunspot data was taken and analysis and preprocessing was done. The comparisons shown is between the Smoothed Data obtained from SIDC website and the smoothening procedure done by using the Savitzky–Golay filter. This was done to confirm the 13 Month smoothening done by the SIDC. 
I then breakdown how the Savitzky-Golay filter works by partitioning each part of the dataset and visualizing the regression done by OLS method. This is how the filter works inside the function and that is how it fits the regression lines based on the polynomials being fitted.